fix lint and add check

+++ metadata-ingestion/src/datahub/cli/delete_cli.py
@@ -198,9 +198,11 @@ def delete(
                 )
             )
             if not force:
-                remove_references = click.confirm("Do you want to delete these references?")
+                remove_references = click.confirm(
+                    "Do you want to delete these references?"
+                )
 
-        if remove_references:
+        if (not force) and remove_references:
             delete_references(urn, dry_run=False, cached_session_host=(session, host))
 
         deletion_result: DeletionResult = delete_one_urn_cmd(
